+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Initial JSON API endpoint for listing deployments
+
 ## [0.3.0] - 2020-09-07
 
 ### Added